requirements.txt
numpy==1.23.1
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
streamlit==1.11.0
tensorflow==2.9.1
